Move ingredients object to its own module
