+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Recipes</title>
     <link rel="stylesheet" href="styles.css">
-    <script src="scripts/script.js" defer></script>
+    <script src="scripts/script.js" type="module"></script>
 </head>
 <body>
     <div class="pagecontent">
