+++ scripts/ingredients.js
@@ -1,0 +1,15 @@
+"use strict";
+
+const tinga = {
+    "onion" : 1,
+    "garlic" : 3,
+    "boxes seitan" : 3,
+    "can tomatoes" : 1,
+    "canned chipotle peppers" : 3,
+    "tsp salt" : 3,
+    "tsp cumin" : 1,
+    "tsp oregano" : 1
+};
+
+
+export { tinga };
