+++ scripts/script.js
@@ -1,16 +1,6 @@
 "use strict";
 
-const tinga = {
-
-    "onion" : 1,
-    "garlic" : 3,
-    "boxes seitan" : 3,
-    "can tomatoes" : 1,
-    "canned chipotle peppers" : 3,
-    "tsp salt" : 3,
-    "tsp cumin" : 1,
-    "tsp oregano" : 1
-};
+import { tinga } from "./ingredients.js";
 
 const recipeObj = {
     "tinga" : {
